Tag 5.4.8-rc1 for internal testing

+++ components/insight/SRC/org/openmicroscopy/shoola/env/ui/OpenObjectLoader.java
@@ -1,6 +1,6 @@
 /*
  *------------------------------------------------------------------------------
- *  Copyright (C) 2006-2015 University of Dundee. All rights reserved.
+ *  Copyright (C) 2006-2018 University of Dundee. All rights reserved.
  *
  *
  *  This program is free software; you can redistribute it and/or modify
@@ -21,6 +21,10 @@
 package org.openmicroscopy.shoola.env.ui;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 
 import org.openmicroscopy.shoola.env.config.Registry;
 
@@ -57,7 +61,37 @@ public class OpenObjectLoader
 
     /** The file where to export the object. */
     private String					folderPath;
-    
+
+    private boolean originalImage;
+
+    /**
+     * Creates a new instance.
+     * 
+     * @param viewer    The viewer this data loader is for.
+     *                  Mustn't be <code>null</code>.
+     * @param registry  Convenience reference for subclasses.
+     * @param ctx The security context.
+     * @param object    The object to handle.
+     * @param folderPath The folder where to copy locally the object.
+     * @param originalImage Download the original image if <code>True</code>
+     *                     otherwise export as OME-TIFF.
+     * @param activity  The activity associated to this loader.
+     */
+    public OpenObjectLoader(UserNotifier viewer, Registry registry,
+            SecurityContext ctx, DataObject object, String folderPath,
+            boolean originalImage, ActivityComponent activity)
+    {
+        super(viewer, registry, ctx, activity);
+        if (object == null)
+            throw new IllegalArgumentException("Object not valid.");
+        if (!(object instanceof ImageData || 
+                object instanceof FileAnnotationData))
+            throw new IllegalArgumentException("Object not valid.");
+        this.object = object;
+        this.folderPath = folderPath;
+        this.originalImage = originalImage;
+    }
+
     /**
      * Creates a new instance.
      * 
@@ -73,14 +107,7 @@ public class OpenObjectLoader
 			SecurityContext ctx, DataObject object, String folderPath,
 			ActivityComponent activity)
 	{
-		super(viewer, registry, ctx, activity);
-		if (object == null)
-			throw new IllegalArgumentException("Object not valid.");
-		if (!(object instanceof ImageData || 
-				object instanceof FileAnnotationData))
-			throw new IllegalArgumentException("Object not valid.");
-		this.object = object;
-		this.folderPath = folderPath;
+		this(viewer, registry, ctx, object, folderPath, false, activity);
 	}
 	
 	/**
@@ -93,14 +120,22 @@ public class OpenObjectLoader
     	File f;
     	if (object instanceof ImageData) {
     		ImageData image = (ImageData) object;
-    		String name = image.getName();
-    		name += image.getName();
-    		name += "_"+image.getId();
-    		path += UIUtilities.replaceNonWordCharacters(name)+"."+OMETIFFFilter.OME_TIFF;
-    		f = new File(path);
-    		f.deleteOnExit();
-    		handle = ivView.exportImageAsOMETiff(ctx, image.getId(), f, null,
-    				this);
+    		if (originalImage) {
+    		    List<DataObject> objects = new ArrayList<DataObject>();
+    		    objects.add(image);
+    		    f = new File(folderPath);
+    		    handle = mhView.loadArchivedImage(ctx, objects, f, false, false,
+    	                false, this);
+    		} else {
+    		    String name = image.getName();
+                name += image.getName();
+                name += "_"+image.getId();
+                path += UIUtilities.replaceNonWordCharacters(name)+"."+OMETIFFFilter.OME_TIFF;
+                f = new File(path);
+                f.deleteOnExit();
+                handle = ivView.exportImageAsOMETiff(ctx, image.getId(), f, null,
+                        this); 
+    		}
     	} else {
     		FileAnnotationData fa = (FileAnnotationData) object;
     		path += UIUtilities.replaceNonWordCharacters(fa.getFileName());
@@ -142,7 +177,23 @@ public class OpenObjectLoader
     public void handleResult(Object result)
     { 
     	if (result == null) onException(MESSAGE_RESULT, null);
-    	else activity.endActivity(result); 
+    	else {
+    	    if (originalImage) {
+    	        Map<Boolean, List<File>> r = (Map<Boolean, List<File>>) result;
+    	        List<File> files = r.get(true);
+    	        if (files.size() > 0) {
+    	            Iterator<File> i = files.iterator();
+    	            while (i.hasNext()) {
+                        i.next().deleteOnExit();
+                    }
+    	            activity.endActivity(files.get(0));
+    	        } else {
+    	            onException(MESSAGE_RESULT, null);
+    	        }
+    	    } else {
+    	        activity.endActivity(result);
+    	    }
+    	}
     }
     
 }
